+++ aksel.nav.no/website/sanity/plugins/aksel-manager/documents/innholdstype.ts
@@ -1,0 +1,38 @@
+import { defineField, defineType } from "sanity";
+
+export default defineType({
+  name: "gp.innholdstype",
+  title: "Innholdstype",
+  type: "document",
+  fields: [
+    defineField({
+      name: "title",
+      title: "Tittel",
+      type: "string",
+      validation: (Rule) => Rule.required(),
+    }),
+    defineField({
+      name: "description",
+      title: "Beskrivelse",
+      type: "text",
+      rows: 3,
+    }),
+  ],
+  orderings: [
+    {
+      title: "Tittel",
+      name: "title",
+      by: [{ field: "title", direction: "asc" }],
+    },
+  ],
+  preview: {
+    select: {
+      title: "title",
+    },
+    prepare({ title }) {
+      return {
+        title,
+      };
+    },
+  },
+});
