+++ aksel.nav.no/website/sanity/plugins/aksel-manager/index.ts
@@ -1,32 +1,25 @@
 import { SANITY_API_VERSION } from "@/sanity/config";
 import { definePlugin } from "sanity";
 import { deskTool } from "sanity/desk";
-import artikkel from "./artikkel";
-import innholdsType from "./innholdstype";
+import artikkel from "./documents/artikkel";
+import innholdsType from "./documents/innholdstype";
+import tema from "./documents/tema";
+import undertema from "./documents/undertema";
 import { defaultDocumentNode, structure } from "./structure";
-import tema from "./tema";
-import temaTag from "./tema_tag";
-/* import skosConcept from './skosConcept'
-import skosConceptScheme from './skosConceptScheme'
-import TreeView from './components/TreeView'
-import {schemeFilter, branchFilter, HierarchyInput} from './helpers'
 
-import {defaultDocumentNode, structure} from './structure'
-import NodeTree from './components/NodeTree' */
+export const GP_DOCUMENTS = [
+  tema.name,
+  undertema.name,
+  innholdsType.name,
+  artikkel.name,
+];
 
-/**
- * Defines a Sanity plugin for managing taxonomies.
- * baseURI should follow an IANA http/s scheme and should terminate with either a / or #.
- * @param options - Optional configuration options for the plugin.
- * @param options.baseUri - The base URI to use for SKOS concepts and concept schemes.
- * @returns A Sanity plugin object.
- */
 const akselManager = definePlugin((options?: any) => {
   return {
     name: "taxonomyManager",
     options,
     schema: {
-      types: [tema(), temaTag(), innholdsType(), artikkel()],
+      types: [tema, undertema, innholdsType, artikkel],
       templates: [
         {
           id: "gp.tema.tag.by.tema",
