--- aksel.nav.no/website/sanity/plugins/aksel-manager/innholdstype.ts
@@ -1,40 +1,0 @@
-import { defineField, defineType } from "sanity";
-
-export default function innholdsType() {
-  return defineType({
-    name: "gp.innholdstype",
-    title: "Innholdstype",
-    type: "document",
-    fields: [
-      defineField({
-        name: "title",
-        title: "Tittel",
-        type: "string",
-        validation: (Rule) => Rule.required(),
-      }),
-      defineField({
-        name: "description",
-        title: "Beskrivelse",
-        type: "text",
-        rows: 3,
-      }),
-    ],
-    orderings: [
-      {
-        title: "Tittel",
-        name: "title",
-        by: [{ field: "title", direction: "asc" }],
-      },
-    ],
-    preview: {
-      select: {
-        title: "title",
-      },
-      prepare({ title }) {
-        return {
-          title,
-        };
-      },
-    },
-  });
-}
