--- aksel.nav.no/website/sanity/plugins/aksel-manager/tema.ts
@@ -1,58 +1,0 @@
-import { defineArrayMember, defineField, defineType } from "sanity";
-
-export default function tema() {
-  return defineType({
-    name: "gp.tema",
-    title: "Tema",
-    type: "document",
-    fields: [
-      defineField({
-        name: "title",
-        title: "Tittel",
-        type: "string",
-        validation: (Rule) => Rule.required(),
-      }),
-      defineField({
-        name: "description",
-        title: "Beskrivelse",
-        type: "text",
-        rows: 3,
-      }),
-      defineField({
-        name: "tags",
-        title: "Tags",
-        type: "array",
-        validation: (Rule) => Rule.unique(),
-        of: [
-          defineArrayMember({
-            type: "reference",
-            to: [{ type: "gp.tema.tag" }],
-            options: {
-              disableNew: true,
-            },
-          }),
-        ],
-        options: {
-          sortable: true,
-        },
-      }),
-    ],
-    orderings: [
-      {
-        title: "Tittel",
-        name: "title",
-        by: [{ field: "title", direction: "asc" }],
-      },
-    ],
-    preview: {
-      select: {
-        title: "title",
-      },
-      prepare({ title }) {
-        return {
-          title,
-        };
-      },
-    },
-  });
-}
