+++ aksel.nav.no/website/sanity/plugins/structure/index.tsx
@@ -23,6 +23,7 @@ import {
 } from "@navikt/aksel-icons";
 import differenceInMonths from "date-fns/differenceInMonths";
 import { StructureResolver } from "sanity/desk";
+import { GP_DOCUMENTS } from "../aksel-manager";
 import { GodPraksisPanes } from "./god-praksis";
 import { Panes } from "./panes";
 
@@ -57,6 +58,7 @@ const filtered = [
   "skrivehjelp",
   "publication_flow",
   "aksel_feedback",
+  ...GP_DOCUMENTS,
 ];
 
 export const structure: StructureResolver = async (
